refactor members models into separate files
