+++ src/involvement/models/__init__.py
@@ -1,9 +1,11 @@
-from involvement.models.application import Application
-from involvement.models.contact_card import ContactCard
-from involvement.models.contact_card_block import ContactCardBlock
-from involvement.models.mandate_history import MandateHistory
-from involvement.models.position import Position
-from involvement.models.recruitment_page import RecruitmentPage
-from involvement.models.reference import Reference
-from involvement.models.role import Role
-from involvement.models.team import Team
+# flake8: noqa
+from .application import Application
+from .contact_card import ContactCard
+from .contact_card_block import ContactCardBlock
+from .mandate_history import MandateHistory
+from .current_mandate import CurrentMandate
+from .position import Position
+from .recruitment_page import RecruitmentPage
+from .reference import Reference
+from .role import Role
+from .team import Team
