+++ src/members/models/__init__.py
@@ -1,0 +1,4 @@
+# flake8: noqa
+from .member import Member, CaseInsensitiveUsernameUserManager
+from .section import Section
+from .study_program import StudyProgram
